requirements.txt
groq
schedule
praw
python-dotenv
